# Reduce the execution time of several unittests